% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCasrnFromNsc.R
\name{getCasrnFromNsc}
\alias{getCasrnFromNsc}
\title{Get CasRN value from an NSC using PubChem PUG API}
\usage{
getCasrnFromNsc(nsc, debug = TRUE)
}
\arguments{
\item{nsc}{an NSC ID}

\item{debug}{a boolean, show debugging information (default: TRUE)}
}
\value{
a CasRN ID or NA, otherwise
}
\description{
Get CasRN value from an NSC using PubChem PUG API
}
\examples{
getCasrnFromNsc("733625")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
