% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChebiFromCasrn.R
\name{getChebiFromCasrn}
\alias{getChebiFromCasrn}
\title{Get Chebi from CASRN using PubChem PUG API}
\usage{
getChebiFromCasrn(id, debug = TRUE, cachePath = NULL)
}
\arguments{
\item{id}{an id}

\item{debug}{a boolean, show debugging information (default: TRUE)}

\item{cachePath}{a path to the cache}
}
\value{
the name used by PubChem
}
\description{
Get Chebi from CASRN using PubChem PUG API
}
\examples{
getChebiFromCasrn("587871-26-9")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
