% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChebiFromName.R
\name{getChebiFromName}
\alias{getChebiFromName}
\title{Get Chebi from Name using PubChem PUG API}
\usage{
getChebiFromName(id, debug = TRUE, cachePath = NULL)
}
\arguments{
\item{debug}{a boolean, show debugging information (default: TRUE)}

\item{cachePath}{a path to the cache}

\item{name}{a name}
}
\value{
the name used by PubChem
}
\description{
Get Chebi from Name using PubChem PUG API
}
\examples{
getChebiFromName("17-AAG")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{
rcellminerPubchem
}
