% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInChiKeyFromCid.R
\name{getInChiKeyFromCid}
\alias{getInChiKeyFromCid}
\title{Get InChiKey from PubChem CID using PubChem PUG API}
\usage{
getInChiKeyFromCid(id, debug = TRUE)
}
\arguments{
\item{id}{an ID}
}
\value{
an InChiKey or NA, otherwise
}
\description{
Get InChiKey from PubChem CID using PubChem PUG API
}
\examples{
getInChiKeyFromCid("1030")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
