% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIsomericSmilesFromCid.R
\name{getIsomericSmilesFromCid}
\alias{getIsomericSmilesFromCid}
\title{Get IsomericSmiles from PubChem CID using PubChem PUG API}
\usage{
getIsomericSmilesFromCid(id, debug = TRUE)
}
\arguments{
\item{id}{a PubChem CID}
}
\value{
a SMILES or NA, otherwise
}
\description{
Get IsomericSmiles from PubChem CID using PubChem PUG API
}
\examples{
getIsomericSmilesFromCid("1030")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{
rcellminerPubchem
}
