% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNameFromSynonyms.R
\name{getNameFromSynonyms}
\alias{getNameFromSynonyms}
\title{Get PubChem Name for a Given Synonym using PubChem PUG API}
\usage{
getNameFromSynonyms(id, debug = TRUE, cachePath = NULL)
}
\arguments{
\item{debug}{a boolean, show debugging information (default: TRUE)}

\item{name}{a name}
}
\value{
the name used by PubChem
}
\description{
Get PubChem Name for a Given Synonym using PubChem PUG API
}
\examples{
getNameFromSynonyms("17-AAG")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
