% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNscFromName.R
\name{getNscFromName}
\alias{getNscFromName}
\title{Get NSC from Name using PubChem PUG API}
\usage{
getNscFromName(id, debug = TRUE, cachePath = NULL)
}
\arguments{
\item{debug}{a boolean, show debugging information (default: TRUE)}

\item{name}{a name}
}
\value{
the name used by PubChem
}
\description{
Get NSC from Name using PubChem PUG API
}
\examples{
getNscFromName("17-AAG")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
