% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNscsFromCasrn.R
\name{getNscsFromCasrn}
\alias{getNscsFromCasrn}
\title{Get NSC value from an CASRN ID using PubChem PUG API}
\usage{
getNscsFromCasrn(casrn, debug = TRUE)
}
\arguments{
\item{debug}{a boolean, show debugging information (default: TRUE)}

\item{nsc}{an NSC ID}
}
\value{
a CasRN ID or NA, otherwise
}
\description{
Get NSC value from an CASRN ID using PubChem PUG API
}
\examples{
getCasrnFromNsc("733625")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{
rcellminerPubchem
}
