% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNscsFromChebi.R
\name{getNscsFromChebi}
\alias{getNscsFromChebi}
\title{Get NSCs from CHEBI ID}
\usage{
getNscsFromChebi(id, debug = TRUE, cachePath = NULL)
}
\arguments{
\item{id}{the ID}

\item{debug}{a boolean, whether to show debug information}

\item{cachePath}{a path for caching results, if not specified the default cache
path for simpleRCache will be used}
}
\value{
returns a data.frame with all CIDs for the CHEBI ID; an annotation 
  of if the CID represents the whole compound "standardized" or a component 
  "component"; and a comma-delimited string of NSCs for the CIDs
}
\description{
Get NSCs from CHEBI ID
}
\details{
"component" compounds could represent things like the ubiquitous HCl 
  or the compound of interest
}
\examples{
getNscsFromChebi("CHEBI:9710")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{
rcellminerPubchem
}
