% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNscsFromCid.R
\name{getNscsFromCid}
\alias{getNscsFromCid}
\title{Get NSCs from PubChem CID}
\usage{
getNscsFromCid(cid, debug = TRUE)
}
\arguments{
\item{cid}{a string with the PubChem CID}

\item{debug}{a boolean, print debug information? (Default: TRUE)}
}
\value{
a vector of NSCs
}
\description{
Get NSCs from PubChem CID
}
\examples{
results <- getNscsFromCid("5284373", FALSE)

}
\concept{rcellminerPubchem}
