% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubchemClassificationFromSid.R
\name{getPubchemClassificationFromSid}
\alias{getPubchemClassificationFromSid}
\title{Determine if PubChem has Use Classifications for a PubChem SID}
\usage{
getPubchemClassificationFromSid(sid, debug = TRUE)
}
\arguments{
\item{sid}{the ID}

\item{debug}{a boolean, whether to show debug information}
}
\value{
returns a vector where the first number is the CID and three additional 
  entries that denote the existence of the following MESH Terms: 
  Pharmacologic Actions: 68020228
}
\description{
Determine if PubChem has Use Classifications for a PubChem SID
}
\examples{
getPubchemClassificationFromSid("53790588")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
