% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubchemFromChembl.R
\name{getPubchemFromChembl}
\alias{getPubchemFromChembl}
\title{Get PubChem SID/CID value from an ChEMBL ID using PubChem PUG API}
\usage{
getPubchemFromChembl(id, type = c("cid", "sid"), debug = TRUE, maxTries = 3)
}
\arguments{
\item{id}{an ChEMBL ID}

\item{type}{"cid" or "sid" PubChem ID types}

\item{debug}{a boolean, print debug information? (Default: TRUE)}

\item{maxTries}{number of times to re-try query}
}
\value{
a PubChem CID/SID or NA, otherwise
}
\description{
Get PubChem SID/CID value from an ChEMBL ID using PubChem PUG API
}
\examples{
getPubchemFromChembl("CHEMBL192", "cid") #Should be: 135398744

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
