% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubchemFromName.R
\name{getPubchemFromName}
\alias{getPubchemFromName}
\title{Get PubChem CID value from a name using PubChem PUG API}
\usage{
getPubchemFromName(id, forceFirst = FALSE, debug = TRUE, cachePath = NULL)
}
\arguments{
\item{id}{a name}

\item{forceFirst}{a boolean, if multiple entries are returned grab the first?}

\item{debug}{a boolean, print debug information? (Default: TRUE)}

\item{cachePath}{path for cache files}
}
\value{
a PubChem CID or NA, otherwise
}
\description{
Get PubChem CID value from a name using PubChem PUG API
}
\examples{
getPubchemFromNsc("trichostatin a") #Should be: 444732
getPubchemFromNsc("aspirin") #Should be: 2244

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
