% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmilesFromCasrn.R
\name{getSmilesFromCasrn}
\alias{getSmilesFromCasrn}
\title{Get Smiles from CASRN using PubChem PUG API}
\usage{
getSmilesFromCasrn(casrn, debug = TRUE)
}
\arguments{
\item{nsc}{an NSC ID}
}
\value{
a PubChem CID/SID or NA, otherwise
}
\description{
Get Smiles from CASRN using PubChem PUG API
}
\examples{
getSmilesFromCasrn("119413-54-6")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
