% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSynonymsFromNsc.R
\name{getSynonymsFromNSC}
\alias{getSynonymsFromNSC}
\title{Get PubChem Synonyms for a Given Name using PubChem PUG API}
\usage{
getSynonymsFromNSC(id, debug = TRUE, maxItems = 10, cachePath = NULL)
}
\arguments{
\item{debug}{a boolean, show debugging information (default: TRUE)}

\item{name}{a name}
}
\value{
the name used by PubChem
}
\description{
Get PubChem Synonyms for a Given Name using PubChem PUG API
}
\examples{
getSynonymsFromNSC("609699")

}
\author{
Augustin Luna (augustin@mail.nih.gov)
}
\concept{rcellminerPubchem}
